Guest list is now being returned and displayed on the dashboard

+++ Paramount.Betterclassifieds.Business/Events/EventBookingTicketField.cs
@@ -2,7 +2,7 @@ namespace Paramount.Betterclassifieds.Business.Events
 {
     public class EventBookingTicketField
     {
-        public int EventBookingTicketFieldId { get; set; }
+        public long EventBookingTicketFieldId { get; set; }
         public int EventBookingTicketId { get; set; }
         public string FieldName { get; set; }
         public string FieldValue { get; set; }
